refactor(django): move the Translatable demo out of the library
django_rakaia shipped a translations model and manager, an admin, an HTMX
dashboard, JSON endpoints, a translations SSE feed and three templates. That is
demo domain, not library surface — the langcode choices were hard-coded to
tet/pt/id — and because the model was declared in 0001_initial, every consumer
got a django_rakaia_translatable table whether they used it or not.

It moves to examples/polyglot, the only thing that ever used it, and whose own
signals.py already made the argument: "we don't decorate the library's
Translatable model itself — that would push demo concerns into the library."

Nothing in the core imported anything from the feature, so this is subtraction
rather than untangling: models.py loses 135 lines, views.py 213, channels_views
66, admin 69, and channels_signals drops the post_save receiver that inspected
translatable_id and broadcast to a hard-coded "translations" group. The one
shared helper, _sanitize_group_name, was core-owned and core-consumed.

Migration 0008 drops the table. UPGRADING.md carries the model definition for
anyone who was relying on it, the dumpdata step, and the removed URL list.

Incidental routing change: a stream literally named `translations` now resolves
to stream_detail, since /streams/translations/ is no longer claimed.

Verified: 1105 tests pass, docs build, all 11 demos pass, and polyglot migrates
and streams a translation end to end on its own model.
