+++ examples/polyglot/polyglot/signals.py
@@ -20,7 +20,7 @@ from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 
 from django_rakaia.decorators import create_stream_event
-from django_rakaia.models import Translatable
+from polyglot.models import Translatable
 
 
 @dataclass
