+++ src/django_rakaia/channels_signals.py
@@ -12,7 +12,7 @@ from channels.layers import get_channel_layer
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-from .models import StreamEntry, StreamEvent
+from .models import StreamEntry
 
 
 def _sanitize_group_name(name: str) -> str:
@@ -32,51 +32,6 @@ def _sanitize_group_name(name: str) -> str:
         c if (c.isascii() and (c.isalnum() or c in "-_.")) else "." for c in name
     )
     return cleaned[-99:]
-
-
-@receiver(post_save, sender=StreamEvent)
-def handle_stream_event_created(sender, instance, created, **kwargs):  # noqa: ARG001
-    """Broadcast translation events to the 'translations' channel group."""
-    # Fixture loads (`loaddata`, `serialized_rollback`) are replayed history,
-    # not live facts — broadcasting them sends phantom frames to connected
-    # clients. See issue #80.
-    if kwargs.get("raw"):
-        return
-    # `data` is any JSON value, not necessarily an object: a protocol append can
-    # carry a list, a string or a number, and the durable store holds a
-    # non-JSON payload as a string. Only an object can be a translation event,
-    # and anything else used to raise `AttributeError` out of `post_save` — i.e.
-    # fail the append, after the write had landed.
-    if not created or not isinstance(instance.data, dict):
-        return
-    if not instance.data.get("translatable_id"):
-        return
-
-    channel_layer = get_channel_layer()
-    if channel_layer is None:
-        return
-
-    message = {
-        "type": "translation.event",
-        "stream": {
-            "id": instance.id,
-            "user": instance.data.get("username", "Unknown"),
-            "langcode": instance.data.get("langcode", ""),
-            "url": instance.data.get("url", ""),
-            "action": instance.event_type,
-            "translatable": {
-                "id": instance.data.get("translatable_id"),
-                "msgid": instance.data.get("msgid", ""),
-                "msgstr": instance.data.get("msgstr", ""),
-                "langcode": instance.data.get("langcode", ""),
-            },
-            "created_at": instance.created_at.isoformat()
-            if instance.created_at
-            else None,
-        },
-    }
-
-    async_to_sync(channel_layer.group_send)("translations", message)
 
 
 def _frame(stream_id: str, entry: StreamEntry) -> dict:
