+++ src/django_rakaia/models.py
@@ -5,22 +5,11 @@ Provides a normalized model structure with Stream, StreamEvent, and StreamEntry
 for efficient querying and real-time updates via Unix sockets.
 """
 
-import enum
-import warnings
-
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
 from django.db.models import Max
-from django.utils import timezone
 
 from rakaia.types import ClosedBy
-
-DEFAULT_LANG = "en"
-
-
-class MSG_IDX(enum.Enum):
-    SINGULAR = 0
-    PLURAL = 1
 
 
 class Stream(models.Model):
@@ -328,142 +317,6 @@ class StreamEntry(models.Model):
 
     def __str__(self) -> str:
         return f"{self.stream.stream_id}#{self.offset}"
-
-
-class TranslatableManager(models.Manager["Translatable"]):
-    @staticmethod
-    def plural_formula(langcode: str, number: int):  # noqa: ARG004
-        """
-        Different languages have different "plural" forms
-        This can be specified here - although most languages have
-        "one thing" , "many thingS"
-        Unless we translate Arabic or Polish we should be good!
-        """
-        if number == 1:
-            return 0  # "an apple", "one apple"
-        return 1  # "no apples", "some apples", "five apples"
-
-    def gettext(self, msgid: str, langcode: str = DEFAULT_LANG):
-        try:
-            return (
-                self.get_queryset()
-                .filter(langcode=langcode, msgid=msgid)
-                .first()
-                .msgstr
-            )
-        except (Translatable.DoesNotExist, IndexError, AttributeError):
-            pass
-        warnings.warn(
-            f"No translated content found: langcode='{langcode}', msgid='{msgid}'",
-            stacklevel=2,
-        )
-
-    def ngettext(
-        self, singular: str, plural: str, number: int, langcode: str = DEFAULT_LANG
-    ):
-        msg_idx = TranslatableManager.plural_formula(langcode, number)
-        try:
-            return (
-                self.get_queryset()
-                .filter(langcode=langcode, msgid=singular)
-                .first()
-                .msgstr[msg_idx]
-            )
-        except (Translatable.DoesNotExist, IndexError, AttributeError):
-            pass
-        warnings.warn(
-            f"No translated content found: langcode='{langcode}', singular='{singular}', plural='{plural}'",
-            stacklevel=2,
-        )
-        return singular if msg_idx == MSG_IDX.SINGULAR.value else plural
-
-    def pgettext(self, context: str, msgid: str, langcode: str = DEFAULT_LANG):
-        try:
-            return (
-                self.get_queryset()
-                .filter(msgctxt=context, langcode=langcode, msgid=msgid)
-                .first()
-                .msgstr
-            )
-        except (Translatable.DoesNotExist, IndexError, AttributeError):
-            pass
-        warnings.warn(
-            f"No translated content found: langcode='{langcode}', context='{context}', msgid='{msgid}'",
-            stacklevel=2,
-        )
-        return msgid
-
-    def npgettext(
-        self,
-        context: str | None,
-        singular: str,
-        plural: str,
-        number: int,
-        langcode: str = DEFAULT_LANG,
-    ):
-        msg_idx = TranslatableManager.plural_formula(langcode, number)
-        try:
-            return (
-                self.get_queryset()
-                .filter(msgctxt=context, langcode=langcode, msgid=singular)
-                .first()
-                .msgstr[msg_idx]
-            )
-        except (Translatable.DoesNotExist, IndexError, AttributeError):
-            pass
-        warnings.warn(
-            f"No translated content found: langcode='{langcode}', context='{context}', singular='{singular}'",
-            stacklevel=2,
-        )
-        return singular if msg_idx == MSG_IDX.SINGULAR.value else plural
-
-
-class Translatable(models.Model):
-    """
-    This represents a database-side interpretation of the `gettext`
-    funtions
-    """
-
-    msgid = models.CharField(
-        max_length=2048, help_text="The original message to be translated"
-    )
-    msgstr = models.CharField(
-        max_length=2048, null=True, blank=True, help_text="Translated message"
-    )
-    domain = models.CharField(
-        max_length=2048, null=True, blank=True, help_text="Message domain"
-    )
-    msgctxt = models.CharField(
-        max_length=2048, null=True, blank=True, help_text="Message context"
-    )
-    langcode = models.CharField(
-        max_length=3,
-        help_text="The destination language code",
-        default=DEFAULT_LANG,
-        choices=[("tet", "tet"), ("pt", "pt"), ("id", "id")],
-    )
-    deleted = models.DateTimeField(null=True, blank=True)
-
-    objects = TranslatableManager()
-
-    class Meta:
-        unique_together = [["msgid", "msgctxt", "langcode"]]
-        indexes = [
-            models.Index(fields=["msgid", "msgctxt", "langcode"]),
-        ]
-
-    def __str__(self):
-        return self.msgid
-
-    def soft_delete(self):
-        """Soft delete by marking as deleted instead of actually deleting."""
-        self.deleted = timezone.now()
-        self.save()
-
-    def restore(self):
-        """Restore a soft-deleted translation."""
-        self.deleted = None
-        self.save()
 
 
 class ConsumerCursor(models.Model):
