--- src/django_rakaia/templates/django_rakaia/_activity_item.html
@@ -1,1 +1,0 @@
-<div class="activity-item"><div class="activity-header"><span class="activity-user">{{ item.user }}</span> <span class="activity-action">{{ item.action }}</span> <span class="activity-lang">[{{ item.langcode }}]</span></div><div class="activity-content"><strong>{{ item.translatable.msgid }}</strong>{% if item.translatable.msgstr %} → {{ item.translatable.msgstr }}{% endif %}</div></div>
