+++ src/django_rakaia/views.py
@@ -5,7 +5,6 @@ Provides a web interface for viewing and monitoring stream events in real-time.
 Uses the normalized Stream/StreamEvent/StreamEntry model structure.
 """
 
-import json
 import logging
 from typing import Any
 
@@ -13,9 +12,9 @@ from django.contrib.auth.decorators import login_required
 from django.db.models import Count, Max, Min
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
-from django.views.decorators.http import require_GET, require_POST
+from django.views.decorators.http import require_GET
 
-from .models import Stream, StreamEntry, StreamEvent, Translatable
+from .models import Stream, StreamEntry, StreamEvent
 
 _log = logging.getLogger("django_rakaia.views")
 
@@ -257,218 +256,3 @@ def streams_api(_request: Any) -> Any:
     return JsonResponse(
         {"streams": serialized_streams, "total": len(serialized_streams)}
     )
-
-
-def _filtered_translations(request: Any) -> Any:
-    """Build the filtered translations queryset shared by HTMX views."""
-    langcode = request.GET.get("langcode") or request.POST.get("langcode_filter") or ""
-    msgid_filter = request.GET.get("msgid") or request.POST.get("msgid_filter") or ""
-
-    query = Translatable.objects.all()
-    if langcode:
-        query = query.filter(langcode=langcode)
-    if msgid_filter:
-        query = query.filter(msgid__icontains=msgid_filter)
-    return query.order_by("msgid", "langcode")[:100]
-
-
-@require_GET
-@login_required
-def translations_table_htmx(request: Any) -> HttpResponse:
-    """Return the rendered <tr> rows for HTMX swap-in."""
-    translations = _filtered_translations(request)
-    return render(
-        request,
-        "django_rakaia/_translation_rows.html",
-        {"translations": translations},
-    )
-
-
-@require_POST
-@login_required
-def translation_create_htmx(request: Any) -> HttpResponse:
-    """Form-encoded create/update via HTMX. Returns the refreshed table body.
-
-    Also creates a StreamEvent so the channels signal handler broadcasts the
-    change to the `translations` group, which the SSE activity feed consumes.
-    """
-    msgid = (request.POST.get("msgid") or "").strip()
-    msgstr = (request.POST.get("msgstr") or "").strip()
-    langcode = request.POST.get("langcode") or "en"
-    domain = request.POST.get("domain") or None
-    msgctxt = request.POST.get("msgctxt") or None
-
-    if not msgid:
-        return HttpResponse("msgid is required", status=400)
-
-    translatable, created = Translatable.objects.update_or_create(
-        msgid=msgid,
-        langcode=langcode,
-        msgctxt=msgctxt or "",
-        defaults={"msgstr": msgstr, "domain": domain},
-    )
-
-    # Emit a StreamEvent so the channels signal broadcasts to the
-    # `translations` group. The HTML SSE view consumes that group.
-    StreamEvent.objects.create(
-        event_type="create" if created else "update",
-        data={
-            "translatable_id": translatable.id,
-            "msgid": translatable.msgid,
-            "msgstr": translatable.msgstr,
-            "langcode": translatable.langcode,
-            "username": request.user.get_username(),
-            "url": request.path,
-        },
-    )
-
-    translations = _filtered_translations(request)
-    return render(
-        request,
-        "django_rakaia/_translation_rows.html",
-        {"translations": translations},
-    )
-
-
-@require_GET
-@login_required
-def translations_index(request: Any) -> Any:
-    """
-    Display translation management interface.
-    """
-    # Get available languages
-    languages = Translatable.objects.values_list("langcode", flat=True).distinct()
-
-    # Get recent translation events
-    recent_streams = StreamEvent.objects.filter(
-        data__has_key="translatable_id"
-    ).order_by(  # Filter for translation events
-        "-created_at"
-    )[:10]
-
-    context = {
-        "languages": list(languages),
-        "recent_streams": [
-            {
-                "id": stream.id,
-                "user": stream.data.get("username", "Unknown"),
-                "langcode": stream.data.get("langcode", ""),
-                "url": stream.data.get("url", ""),
-                "translatable": {
-                    "msgid": stream.data.get("msgid", ""),
-                    "msgstr": stream.data.get("msgstr", ""),
-                },
-                "action": stream.event_type,
-                "created_at": stream.created_at.isoformat()
-                if stream.created_at
-                else None,
-            }
-            for stream in recent_streams
-        ],
-    }
-
-    return render(request, "django_rakaia/translations_index.html", context)
-
-
-@require_GET
-@login_required
-def translations_api(request: Any) -> Any:
-    """
-    API endpoint to get all translations with filtering options.
-    """
-    langcode = request.GET.get("langcode")
-    msgid_filter = request.GET.get("msgid")
-
-    query = Translatable.objects.all()
-
-    if langcode:
-        query = query.filter(langcode=langcode)
-    if msgid_filter:
-        query = query.filter(msgid__icontains=msgid_filter)
-
-    translations = query.order_by("msgid", "langcode")[:100]
-
-    serialized = [
-        {
-            "id": t.id,
-            "msgid": t.msgid,
-            "msgstr": t.msgstr,
-            "langcode": t.langcode,
-            "domain": t.domain,
-            "msgctxt": t.msgctxt,
-            "deleted": t.deleted.isoformat() if t.deleted else None,
-        }
-        for t in translations
-    ]
-
-    return JsonResponse(
-        {
-            "translations": serialized,
-            "total": len(serialized),
-            "languages": list(
-                Translatable.objects.values_list("langcode", flat=True).distinct()
-            ),
-        }
-    )
-
-
-@require_POST
-@login_required
-def translation_create_update_api(request: Any) -> Any:
-    """
-    API endpoint to create or update translations.
-    """
-    try:
-        data = json.loads(request.body)
-        msgid = data.get("msgid")
-        msgstr = data.get("msgstr")
-        langcode = data.get("langcode", "en")
-        domain = data.get("domain")
-        msgctxt = data.get("msgctxt")
-
-        if not msgid:
-            return JsonResponse({"error": "msgid is required"}, status=400)
-
-        # Get or create translation
-        translatable, created = Translatable.objects.get_or_create(
-            msgid=msgid,
-            langcode=langcode,
-            msgctxt=msgctxt or "",
-            defaults={
-                "msgstr": msgstr,
-                "domain": domain,
-            },
-        )
-
-        if not created:
-            translatable.msgstr = msgstr
-            if domain:
-                translatable.domain = domain
-            translatable.save()
-            action = "update"
-        else:
-            action = "create"
-
-        # The Translatable.save() method automatically creates StreamEvent and StreamEntry
-        # No need to manually create TranslationStream anymore
-
-        return JsonResponse(
-            {
-                "success": True,
-                "translation": {
-                    "id": translatable.id,
-                    "msgid": translatable.msgid,
-                    "msgstr": translatable.msgstr,
-                    "langcode": translatable.langcode,
-                    "domain": translatable.domain,
-                    "msgctxt": translatable.msgctxt,
-                },
-                "action": action,
-            }
-        )
-
-    except json.JSONDecodeError:
-        return JsonResponse({"error": "Invalid JSON"}, status=400)
-    except Exception:
-        _log.exception("translation_create_update_api failed")
-        return JsonResponse({"error": "Internal server error"}, status=500)
