+++ zensical.toml
@@ -36,7 +36,7 @@ nav = [
     { "Close preconditions (spike)" = "close-preconditions.md" },
     { "Multi-stream merge (spike)" = "multi-stream-merge.md" },
     { "Tree-reconcile (spike)" = "tree-reconcile.md" },
-    { "Translations" = "translations.md" },
+    { "Translations (example)" = "translations.md" },
     { "Deployment" = "deployment.md" },
     { Reference = [
         { "Protocol specification" = "protocol.md" },
